Scope Improvements
- The bucketing process for the spectrum analyzer now uses a non-linear
progression of sample buckets.  Lower frequency buckets contain more
samples, and upper frequency buckets get less samples.  This ensures
that the graph shows the interesting detail of the lower frequencies
without hiding the higher.  This is currently a compile time option, but
in future could be turned on or off.

- The graph now draws without gaps between the buckets if the window is
too small; but there is an optional value to enable this.  I think it looks
better this way, and even on a bigger window you might not want the
gaps.  The gaps are more suggestive of frequency 'buckets' though.

- Added a thread which collects the audio data from Supercollider and runs
the FFT.  This offloads the work from the GUI.  The thread collects its
data and asks the gui to repaint.  It will not provide a new set of data
until the gui has consumed it, to save flooding the GUI with repaint
requests.  This threading change obviously needs thorough testing.

- Add the tracy analyzer for performance work. Only used on windows,
with debug and release + PDB builds.