Add starting and stopping xcompmgr on Raspberry Pi
On Raspbian Buster you need to start xcompmgr so that Qt commands to vary the opacity of the main window work.
The binary is executed as the server is started, and is closed again at exit. There is an instance running already set up at boot time but it odes not seem to be accessible to Qt. Hence this needs to be done. There are some error messages from xcompmgr but they don't affect the operation required and are sent to /dev/null. The mods have been checked on macOS to make sure they are not invoked.