Upgrade fast_osc to v1.2.0
This commit bumps the version in the vendor folder.

It also adds rake-compiler to the vendored gems to allow compilation to
work for fast_osc. The rake task has to be called from within
`compile-extensions.rb` which looks a little crazy but it's fine,
honest.

It should rescue errors when compiling fast_osc specifically, as it will
fallback to the pure Ruby code in that gem instead so users won't notice
a difference. Other vendored gems with C extensions are unaffected by
this change.

For Windows it also attempts to include `devkit` as a _rake_ dependency
which allows the rake compile task to work, assuming you used Ruby
Installer to install Ruby on your Windows, and that you setup the MSYS2
toolchain (which happens by default).

In short, this shouldn't cause any painful changes and the new features
in fast_osc (potential support for OSC bundles and array args for
supercollider) will work transparently.