udate jackd boot for RPi and Linux
Latest Buster on RPi uses differe .asoundrc Updated check for soundcard to use pcm.output instead of pcm.!default.
Also for both RPi and Linux the wrong pid value was stored, and need to use spawn exec... to get the right one.Similar to commit #1976 which corrected this for m2o and o2m boot in studio.rb
Also connected input ports for RPi, so they work if there is an input device. Error in log if not, but still runs.

I've tested the RPi changes, but don't have suitable setup to test Linux changes, if someone else can check that.