GUI - improve formatting of info tabs for high DPI monitors
Specify the size of images in the html sections in 'dx' (dpi scaled pixels) and scale these values based on the monitor DPI when generating the info widget. This is a similar trick to that used for scaling values in the stylesheets.

The scaling mechanism (a naive search and replace) should ideally be generalised and used by the stylesheet scaling too.