GUI - re-enable shortcuts for zooming log and cue panes
Using stylesheets to set the font size of the cue and log panes inhibited the zooming shortcuts from working.

This is because we set the font size in the global stylesheet selector

*
 {
  font-size: medium;
}

We have now removed this, set the font size in all the appropriate selectors explicitly and set the font size of the log to an appropriate size on creation.

Shortcuts for zooming the cue and log now work again :-)