Scripts - fix clear-logs.rb on Windows
Previously we attempted to create a history directory using the current Timestamp. However, the default text version of a Timestamp in Ruby contains characters not permitted for directory names on Windows. This script therefore silently failed and the logs were never cleared between application launches.

We now sanitise the new folder name replacing any non-valid characters with underscores.

The logs are now correctly cleared on Windows.