Core - remove use of Atom and Hamster Hash
This code was written just as I was moving from Clojure back to Ruby and I needed Clojure-like concurrency semantics to build something concurrent that worked.

This stuff works great in Clojure, but it's not particularly idiomatic over here, so switch back to something more familiar - plain Hash and Monitor.

Note that we need to use a Monitor as it needs to be re-entrant: kill_job is called from a block passed to each_id. This wouldn't work with a basic Mutex as the nested synchronize would block.