Core - Implement own immutable vec and hash
This is a first pass and there's likely to be issues and room for improvement. These are now based on encapsulation of frozen Core classes rather than Hamster's classes.

This is the last part of the codebase that uses Hamster, we are now free to remove it as a dependency.