Runtime - temporarily patch memory leak (partially)
There appears to be a memory leak in in_thread.

I've not managed to track it down yet, so this patch removes a very frequent call to in_thread and switches it over to a regular thread. This has the minor negative impact that it won't be automatically killed on stop, so long schedule ahead times will mean that even when you press stop, the message may still appear afterwards.

This really should be sent through the Erlang router anyway, which has both support for delaying messages and auto flushing unsent messages on stop. A task for the future.

Also, the in_thread memory leak needs to be hunted down and squashed!

Helps to mitigate #2168