Runtime - rework subthread waiting
In a recent commit a memory leak within in_thread was fixed.

This unfortunately then unmasked a race condition issue with how subthreads are both manipulated (added and removed) and used to wait until a thread and its sub children have completed.

The issue was that the waiting essentially iterated of a collection which was being mutated by other threads. Not good. This was masked by the memory leak as the error that was causing the leak was because the collection wasn't being correctly mutated. No mutation = no issues ;-)

We now no longer iterate over all the subthreads rather use a Promise to communicate back when all the subthreads have completed. This can happen either when a thread has completed and it has no subthreads, or when a subthread completes and its parent is completed and it is the sole thread in the parent's subthread list.

Mixed into this commit is a bit of refactoring of when thread locals are initialised. We now use the same init fn for both job threads and in_thread threads.