Use os:system_time() instead of erlang:system_time()
This ensures that the timestamps are better synced with Ruby and others,
avoiding cases where a message appears received slightly before it was sent.
The erlang:system_time() function guarantees monotonic time and also tries to
avoid sudden jumps in time if the OS clock jumps. Sonic Pi however just wants
the current OS clock as precisely as possible.