Fix for crash when hitting WIN+/
This triggers some kind of special mode in the Qt Scintilla layer.
Short story is that it was converting a 64 bit pointer to a 32 bit long,
meaning that the memory at the far end of the message was invalid.

The rest of the code calls this GETTEXT message without the cast to
sptr_t (which is the problem), so I've done the same by casting to (void*)