Generate noise from Python script - add light_pink, dark_pink too
Included one-off Python script `generate_noise.py` to show how the noise
was produced.

Pink noise is 1/f noise with an exponent of 1 - An exponent of 0.5 is
halfway between pink and white noise. I'm going to propose the
non-scientific name :light_pink for that one (more random than pink).
Similarly exponent of 1.5 is halfway between pink noise and brownian
motion (e.g. a random walk) so I'd suggest :dark_pink for that.

The Perlin noise produced by the previous script was quite basic and
uninteresting. To get more varied textures, a different Python script
was used which introduces the idea of octaves to create more detail and
an unbias function to get values away from oscillating around the
midpoint of 0.5. This is all explained in detail here:
https://eev.ee/blog/2016/05/29/perlin-noise/