GUI - show slide opts table for sample docs
Previously, although some opts of the StereoPlayer synth were marked as
slidable, the table in the documentation that described slide opts (which is
visible for synths where relevant) was not visible for the samples
documentation.

This change makes the slide table visible in the samples documentation, since
the StereoPlayer does indeed have slidable opts, as this seems more consistent
with the synths documentation.