Core - fix incorrect error line reporting
The error line detection system was assuming that the implicit thread that each run is executed in was added to the code string to be evalled. This has been changed to a call to in_thread and the string was therefore no longer changed and the offset no longer needed to be modified to take this into account.