OSC - teach encoder how to encode binary blob arguments
Introduced new type SonicPi::OSC::Blob with which to encapsulate a string with binary encoding. This allows the arg encoder to correctly dispatch encoding mechanism based on arg type (Strings are already encoded with a specific format, so we need to distinguish between standard strings and strings containing binary).

Decoding binary blob args was already implemented.