External - add sp_midi Erlang nif
This will eventually replace osmid (o2m and m2o).

sp_midi is developed by Luis Lloret (the author of osmid) and the development can be tracked here: https://github.com/llloret/sp_midi

Instead of being stand-alone binaries sp_midi works within the Erlang process as a nif. This will reduce the number of processes to manage, reduce the number of ports to open, decrease latency (no need to use the loopback network) and increase control as we can now use the features of osmid directly as Erlang functions.

There will be a bit of turbulence as we move across, but hopefully disruption will be kept to a minimum. The aim is to move to sp_midi and remove osmid by 3.3.