GUI - fix weird quirk in QPushButton on macOS
With the latest Qt on macOS (5.15) setting a qss style entry for QPushButton appears to render the button unclickable . Having no stylesheet entry appears to fix things (although without styling) and setting the button to be flat with `setFlat(true) appears to fix things.

So let's try setting the button to be flat. Hopefully this doesn't negatively affect other platforms....