MIDI - reset whole Erlang BEAM when resetting MIDI
Unfortunately there appears to be a bug in JUCE on macOS which stops the MIDI port lists from being updated whilst the NIF is loaded.

We now have to tear down the whole Erlang BEAM and restart it to see changes in the MIDI port listings.

This is brutal and hopefully things will be fixed by JUCE 6 out in the next few weeks or so.

This only affects macOS, so I'll remove the Reset MIDI button on other platforms as it's not necessary.