Docs - update synth design document
The previous update to this file was made prior to some folders in the
repository being moved around, but the update of this file was merged *after*.
As such, some of the file paths described in this document were outdated.

These outdated file paths have now been fixed. Additionally, some sections of
this document have been updated and expanded to provide extra information, and
in a way that possibly reads better.