doc: Update linux install guide
Recently the script files were renamed from unix-* to linux-* so at the
very least this needs to be captured. Secondly, it seems there's two
additional dependencies for erlang required (erlang-dev and
erlang-tools).

While in there, I decided to tidy up some of the text level formatting
to make it easier to read as a text file but also for future
maintenance. One example of the later is removing the table detailing
how to install the additional dependencies. This was done to facilitate
easily seeing what's in the list, and adding new ones as needed - such
as the above mentioned erlang ones.