I18n - give list items in translated files bullets/numbers
Previously, the i18n-tool.rb script may have already been doing this and then
stopped working at some point, or it may never have worked, I'm not sure.

The reason it was not working prior to this change is that it was assumed, when
processing 'p' elements, that a variable ('bullet') may be present containing a
bullet or number to assign to this 'p' as a list item child.
However, no such variable value was ever passed through to the 'p' processing
code - it was always a value of 'nil', so bullets or numbers were never added to
the translated text.

Now, however, it will correctly output list items with bullets or numbers
as appropriate when creating translated files from the English versions.