MIDI - fix bug with some fns with on: 0
If the "truthy"ness test fails for the fn's on: opt, then it shouldn't trigger yet should still print a message to the log (sayign it's a rest or equivalent).

In some cases, the non-trigger message contains information about the fn call (such as target port, channel etc) that was only bound within the true clause of the truthyness test and therefore was not available to the non-trigger message.

These should now all be addressed and calling each fn with on: 0 should not raise any errors.