Core - fix bug in take_last
As first raised on in_thread at
https://in-thread.sonic-pi.net/t/a-drum-n-bass-genre-study/1086/14, (thanks to
https://github.com/nlebellier :) ) - previously, the range of values calculated
to grab out of the ring that was passed the `take_last` message was returning
nil for the min and max of the range. This is because take_last was using -1 for
the range max - which means that since the max was less than the min, nil was
returned for these values.
Now, we use size-1 for the range max, so that it will be larger than min, and
the min and max values will not be nil.