GUI - ensure that  the visibility of buttons, tabs & log are saved/restored in settings
This was previously hard coded to be forced to be visible on app start way back when people got confused that they were somehow hidden. Unfortunately this behaviour got badly mixed with the settings logic so that the check boxes could get into a state where they didn't reflect that actual visibility of the elements. This is now fixed.

In the future we should add the ability to toggle the visibility of these elements from the menu bar for even further increased accessibility and discoverability of this functionality.