Build - update macOS build instructions
We no longer need to explicitly install SuperCollider or Erlang as they are now included as prebuilt binaries on the tree. Also the version of Ruby shipping with Catalina works just fine so there's no need to install the latest Ruby just for the purposes of building on user's personal machines (however, the releases will likely continue to ship with their own internal Ruby tracking the latest releases).