Better audio thread performance
I recently learned that thread::yield does not yield the rest of the time slice; it can result in a spinning thread (which I've seen), since the thread can be immediately rescheduled.
In this case, if the UI hasn't caught up, we want to wait for the next time slice.
A quick comparison on my local machine gave up to .5% less CPU time when running the same sample with FFT on.

Needs testing to confirm nothing is amiss, but looks good.