GUI - fix visibility of "unsupported" syntax in high contrast mode
Note that it should be stated that this isn't technically a bug as using syntax such as Ruby-style @accessors is absolutely not supported and only works because Sonic Pi is currently built on top of Ruby.

There are long-distance plans to consider other host languages and if and when we make this change, the code using @accessors with the @ sigil will no longer work as you expect.

As a guide, we only support the subset of Ruby (and our own extensions) that are included in the tutorial. Using anything else may or may not work and whilst you're absolutely allowed to do it, it's not supported in any way, may break in the future with no warning and may not work as expected.

Fixes #2545