Lang - don't panic if promise has already been delivered.
In this case it's entirely plausible that multiple GC threads could attempt to trigger the same message (i.e. deliver the same promise). The promise implementation guarantees that the message will only be delivered once anyway. It's just that in this case it's not an error to deliver it more than once.