GUI - start compton on RPi for GUI transparency
This currently happens as part of the Ruby Server but probably should take place in the GUI as it is not a requirement or dependency of the server and may get in the way of any future plans for a headless server.

Therefore consider moving this to the GUI in the future.