Boot - don't store pid commands on macOS
When executing within a hardened runtime on macOS it appears that calling `ps` either directly or via `Sys::ProcTable.ps` is denied causing the Ruby process to be immediately terminated.

We called `ps` in order to get the command name of the pid we were registering to store it alongside and to check with when attempting to clear and kill running processes.

We no longer do this on macOS and therefore avoid the Ruby process from being terminated prematurely.

Eventually we need to look into re-designing the process creation and management system from a suite of Ruby scripts to a reliable running process that we connect to via a TCP connection which when disconnected will cause all registered pids to be terminated.