Sample Onsets - move from Ruby FFI to external binary
Previously we called the sample onset detection functionality in the aubio library via a Ruby -> C FFI bridge. This approach unfortunately appears to cause problems when executing in a "hardened runtime" on macOS Big Sur. This patch moves this FFI-based Ruby gem to a static binary which uses stdout to return the onset values.

Onset times should remain identical as the binary has been tuned in the same way as the original Ruby gem:

# [:window_size]     1024
# [:hop_size]        512
# [:onset_threshold] 0.3
# [:minioi_ms]       12.0 (ms)