API
This is a substantial change to the connection between the Qt GUI and
the server.  A new API library acts as the go between.  This CL also
introduces a simple test UI written in ImGui as a client to the API.

The API is a static library and is currently responsible for:
- Finding and initializing paths for logs, etc.
- Finding and checking ports
- Running various ruby scripts, including the server
- Connecting to the supercollider audio api.
- Handling incoming osc messages
- Detecting startup and shutdown errors
- Shutting down the server and osc handler

The API does not yet codify all the outward communication to the server,
but it does provide a SendOSC method.  Eventually it will have type-safe
APIs for all communication to the server. Some APIs such as
SaveAndRunBuffer, and SaveWorkspaces are implemented, and form the
template for the rest of them.

The ImGui API has:
- A simple editor window, loaded with workspaces
- A scope showing the the spectrum analysis
- Windows for logs, midi and cues
- The ability to run/stop the current text buffer
- Shortcut keys ALT+R/ALT+S and a buffer flash to show evaluation

The GUI runs best on Windows, but works on Mac/Linux too (it needs some
further work for DPI/scaling and window loading support)

There is also now a simple unit test application which is run as a
github action.  This is capable of starting a headless version of
SonicPi and playing a note on a virtual audio device.  It ensures that
on all 3 platforms the API can get to the point where Sonic Pi has been
initialized and is ready for playing music.

This CL swaps over to using the API from the Qt Gui, but it can work the
old way for now by uncommenting QT_OLD_API in the root CMakeLists file
(the API and API Gui are always built, this just changes how the main
Gui works).

The new code contains a streamlined version of the Scope window without
all the thread extras, and some more cleanup.

- The api uses vcpkg for package management; it is a better system for
C++ libraries.  It has been tested on Mac/Windows/Linux

The sensitve changes in this CL are:

- The process management uses a library called reproc; because the API
no longer links to Qt.  This means that the process loading code is
different, although it follows the same flows as before.
- The management of the supercollider audio callback is different.