Runtime - initialise rand distribution of new threads
Previously, the thread local tracking the current thread's random number
distribution was only set to a specific value if the user explicitly called
use_/with_random_source.
(Otherwise, a white noise distribution was used _without setting the
corresponding thread local_).
With the introduction of a function to query the corrent random source,
returning a source of nil is undesirable - so here, when creating a new thread,
we query the thread local for the current random number distribution - _and
explicitly set it to :white if it is nil_.