Docs - Fix rand_look and rand_i_look examples
Previously, the behaviour of the last two lines in the affected examples was
reversed. (Ie, rand and rand_i would *not* in fact print new random numbers, but
the same ones as the preceding calls to rand_look and rand_i_look, since
rand_look and rand_i_look 'look ahead' at what the calls to rand and rand_i
would produce, and the subsequent calls to rand and rand_i will advance the
random stream, meaning that calls to rand_look and rand_i_look after this will
look at a *new* random number.

Fixes #2693