Erlang - make UDP sending more robust
Previously the function to send UDP just used gen_udp:send which calls exit if the hostname can't be resolved - killing the process.

As the hostname comes from the code, this could easily be set to the wrong value by the user. Therefore, rather than assume it's a valid error and exit as you should in Erlang, we first check the hostname. If it can be resolved we send it (still via gen_udp so some of these checks are duplicated) if not, we write a debug log message and ignore it.

This should improve the robustness of the Erlang server in the case of these bad host names.