Server - update doc generating script & templates
The translation pot file needs to contain references to files from *all* of
the different documentation sections, to reduce duplication. To achieve this, we
now process all of the collections of documentation data (synths/fx/lang/etc) as
a single group.

Several minor updates to the base templates have also been made.