Sample Onsets - include final sample slice in onset_slices
Previously, when generating the list of sample onset slices, all but the last
slice were being included in the list. This has now been fixed.
The fix also revealed a further problem: the error raised for invalid onset
procs referred to a non-existent variable and therefore itself caused an error.
This also has been fixed.

Possible TODO: handle float precision for onsets that are detected at the very
end of the sample?