Introduce the Boot Daemon
Refactor code that boots long living processes into one central location - a new Boot Daemon process.

This means that the GUI (or actually the API) only needs to boot one process - the Boot Daemon which will take care of launching all long-living processes and ensuring they are appropriately stopped and cleaned up either on successful termination of the app or in the case of a crash.

As part of this refactor, many of the ruby scripts in app/server/ruby/bin directory are no longer neccessary. We no longer store the pids of the long living processes in a file - instead this is stored in memory within the Boot Daemon. We therefore no longer need to register, store, clear these processes. Not only was this error-prone in some crash scenarios, it also became problematic on macOS due to recent security restrictions limiting the ability to kill other processes (which clearly makes a lot of sense). We now only kill processes we spawned which is much more friendly.

Additionally, we now delay any comms from the Spider Ruby Server to the Tau Erlang Server until Tau is up and running (and responds to a /ping). This happens through a new TauComms class which is largely a thread-safe wrapper around UDPClient (with some additional code to wait for Tau to be ready).

As an added bonus, boot times now appear to be reduced.

For more information about working with the new Daemon Booter, here's a copy of the inline comments:

```
# The Sonic Pi Boot Daemon.
# =========================
#
# Process Spawning
# ----------------
#
# The Daemon launches and watches over the following
# long-living processes (necessary for Sonic Pi to work):
#
#   +--> Spider  - the Ruby Runtime server
#   +--> Tau     - the Erlang IO server
#   +--> Scsynth - the SuperCollider audio engine
#
# The Daemon does all the work necessary to figure out the correct
# process paths and flags - even considering config files such as
# `audio-settings.toml`
#
#
# Zombie Kill Switch
# ------------------
#
# It is the Daemon's responsibility to ensure that the processes it
# boots are always terminated and are not left to turn and stick around
# as zombies (processes still accidentally running on your computer
# consuming resources).
#
# This is achieved via a "Zombie Kill Switch" - a TCP connection with
# the an external process (such as the GUI) to monitor its status by
# continually receiving keep_alive messages. If these messages stop
# being received (for example, if the GUI process exited normally or
# even crashed) then the Daemon will ensure all the processes it spawned
# (Spider, Tau and Scsynth) are terminated.
#
# The port number of this kill switch TCP connection is printed to
# STDOUT. A external process (such as the GUI) must connect promptly and
# periodicaly send some bytes (more frequently than every 3s) otherwise
# a timeout will trigger the kill switch and kill all the spawned
# processes.
#
#
# Port Allocation
# ---------------
#
# Boot Daemon figures out appropriate (and currently free) values for
# all the ports used by various processess within Sonic Pi to
# communicate with each other. These ports are used to create the
# correct process arguments for spawning and are also passed to STDOUT.
#
# Some of these port numbers need to be known by external processes -
# such as a GUI process sends code to run and receives log updates both
# via UDP to specific ports.
#
# The current allocations of these external port numbers are printed to
# STDOUT the current allocations for the following order:
#
# daemon-keep-alive gui-listen-to-server gui-send-to-server scsynth osc-cues
#
#
# External Port Descriptions
# --------------------------
#
# daemon-keep-alive:    TCP port Daemon uses to communicate with GUI (or
#                       other controlling process)
#
# gui-listen-to-server: UDP port which the GUI uses to listen to messages
#                       from the Spider Server.
#
# gui-send-to-server:   UDP port which the GUI uses to send messages to
#                       the Spider Server.
#
# scsynth:              UDP port on which scsynth listens (necessary for
#                       connecting to the boost shared memory for scope data)
#
# osc-cues:             UDP port used to receive OSC cue messages from external
#                       processes.
```