Daemon - improve exit routine
It turns out that calling exit from another thread whilst an at_exit block is executing can cause it to stop before completion.

The whole point of the Daemon Booter is that the at_exit block is *always* executed to completion.

To improve the chances of this, we now have a thread safe exit class which ensures that the at_exit block isn't interupted, executes only once and if it's already in the process of running, further calls to the exit method will just kill the current thread and *not* call Kernel.exit.

Kernel.exit should only be called within the SafeExit class and nowhere else for this to behave correctly. To call exit we now need to use SafeExit#exit.