Spider - further work reducing surface area of time functions
All Time Management functions have now been gathered and grouped to a small section at the top of runtime.rb.

Note that all the thread locals in this block of fns are *only used within this fns* and not elsewhere. This includes both reading (get) and writing (set).

This should with the integration of Link support and for a global metronome.