Update view templates
The variable tracking the side panel navigation headings has been renamed, to
more accurately reflect what it represents.

We now show a basic title or heading on each page in each of the main
categories.

Lastly, the 'Tutorial' and 'Examples' categories have been disabled as adding pages to
those categories requires more work, which will not be done unless future
discussions confirm this Phoenix app as the right approach.