Vendor - upgrade ruby gettext
The reason for this change is to allow for some potential work towards
rebuilding the documentation system.

There are several ways the documentation system might be rebuilt - one of them
appeared to require a more recent version of gettext (for reasons I cannot
unfortunately recall at the moment).

In the case that that particular approach to the new documentation system is not
used, this change will not strictly be necessary, *but* there also appear to be no
breaking changes in gettext since the current version, so theoretically this
change should be harmless either way.