Boot - create comms between Daemon and Tau
This enables Tauto send the system pid of the erlang beam process over a TCP connection to Daemon now that the Elixir release scripts boots things indirectly.

The comms channel consists of a single TCP connection which expects OSC in a similar manner to SuperCollider - 32 bit unsigned big endian data size followed by that many bytes of OSC message.

This new TCP socket now also acts as the mechanism for the Daemon to determine whether or not Tau is still alive.