Prebuild - overwrite mix releases by default
Previously, when the Elixir/Erlang artifacts have been already built by a previous build
of Sonic Pi, and we run the Sonic Pi build process again, mix asked whether we
wanted to overwrite them or not. We now explicitly indicate that we want to
overwrite on the command line to avoid being asked during the build process.