Tau - add killswitch
Now that we have changed Tau's boot scripts to use the release scripts from Elixir, we no longer can get the OS pid from the spawn process. The pid is necessary so that we can explicitly kill Tau when we shutdown or something goes wrong. We attempted to circumnavigate this by explicitly passign the pid from Tau to Daemon via a TCP socket. However, there are still some edge cases where this pid isn't correctly sent/received. To mitigate this and to add extra reassurances that Tau will properly exit and not turn into a zombie process we now have an internal kill switch within Tau. This is a simple timer which when fired will bring the whole process down. The timer is reset by the Daemon sending an appropriate keep alive OSC message. If no keepalive message is received within 5 seconds Tau will shut itself down.