Tau - print log directly to file rather than stdout
We now pass the log file path to Tau on boot which then uses it to directly pipe the logs towards. This is instead of the previous approach which piped stdout from Tau to Daemon and then Daemon wrote it to the log file.

This appears to fix an issue with the kill switch in the case where the Daemon process was killed. In certain circumstances (observed on both Windows and macOS) it was possible to block the kill switch from working by killing the Daemon which was working with the stdout. It's possible the BEAM can't detect when something consuming stdout has disappeared and therefore blocks.

Printing to a file rather than stdout appears to stop this blocking from happening and the kill switch seems to work as expected regardless of how the Daemon process is terminated.