Tau - improve log handling
On macOS and Linux close stdout, stderr and stdin which have been inherited by the calling process (likely Daemon) before opening our own pointing to the tau log path. This reduces chances from the tau kill switch from hanging on these platforms.

We also need to move our config to the file runtime.exs for the shell variables to be available to configure the logger.