Docs - Show needed buffer: opt in record fx usage example
Previously, usage examples for all fx in the docs had the same format:

with_fx :fx_name do
  play 50
end

However, the :record fx requires a buffer: opt with a value representing a
buffer for the fx to record into. The existing usage example format does not
account for this.

We have now added a simple condition into the code that builds the fx usage
examples, to add in the required buffer opt if the example is for the :record fx.