Tau - rename midi/link_enabled to midi/link_on
midi_enabled is a potentially confusing name as it's possible to think that this is enabling or disabling the entire feature whereas it actually just means whether that feature is on or off and can be toggled during runtime.

This is the same for link_enabled too.

Further confusion is introduced with the ENV VAR TAU_DISABLE_MIDI which does disable the MIDI functionality entirely at boot time.