Tau - allow booting without MIDI or Link functionality
It is now possible to boot and use Tau without having compiled either of the MIDI or Link NIFs.

By default, neither NIF is loaded, their respective event servers are not created and all NIF functionality is a null op.

However, if the env vars TAU_MIDI_ENABLED and TAU_LINK_ENABLED are set to true then functionality is enabled. These env vars are set to true as part of the standard Sonic Pi boot process.

The intention of this is to make it easier to work on the Elixir/Erlang part of the code base without having to compile the NIFs and also to potentially use/deploy to contexts where these NIFs don't make sense to have enabled (such as a cloud server).