Tau - fix logging
Logs are back up and running - also now defaults to a new log dir if the env var TAU_LOG_PATH isn't set.

All stdout and stderr is sent to a new log file in tau/log/tau_stdout.log. It's assumed that this file will be empty during normal operation and all logging is done via the system logger and is only there for dev work.