Tau - no longer mess with stdin/out/err on boot
This was an attempt to find a way to make sure the daemon correctly shut things down if tau crashed in a specific way which was seemingly blocked by io pipes with child processes.

I couldn't seem to fix it using this approach, so there's no need for this exec magic stuff any more.