Tau - add Phoenix
The Phoenix web server is now integrated into Tau and started by default.

It runs on a randomly assigned port and currently just serves the default Phoenix boiler plate HTML and runs in production mode via the _build release built as part of the prebuild script.

However, you can both configure it to run in dev mode (and therefore support hot loading code changes) and specify the port using a new config file which you need to explicitly create in ~/.sonic-pi/config/tau-settings.toml. Currently the only supported values are env and phx_port (more options may be added in the future). An example file is as follows:

env = "dev"
phx_port = 8888

The C++ API now also has access to the port Phoenix is running on via the key phx_http.

It should also be possible to work directly with the Phoenix part of the code base via mix using the standard processes without having to compile either the MIDI or Link NIFs.